APIからrepositoryを剥がした
