+++ application/api/entity/sns/ios_voip_push_sns_message.go
@@ -1,7 +1,8 @@
 package sns
 
-type IOSVoIPPushSNSMessage struct {
-	CharaID      string `json:"charaID"`
-	CharaName    string `json:"charaName"`
-	VoiceFileURL string `json:"voiceFileURL"`
-}
+//
+//type IOSVoIPPushSNSMessage struct {
+//	CharaID      string `json:"charaID"`
+//	CharaName    string `json:"charaName"`
+//	VoiceFileURL string `json:"voiceFileURL"`
+//}
