+++ application/api/entity/sqs/ios_voip_push_alarm_info_sqs_message.go
@@ -1,10 +1,11 @@
 package sqs
 
-type IOSVoIPPushAlarmInfoSQSMessage struct {
-	AlarmID        string `json:"alarmID"`
-	UserID         string `json:"userID"`
-	SNSEndpointArn string `json:"snsEndpointArn"`
-	CharaID        string `json:"charaID"`
-	CharaName      string `json:"charaName"`
-	VoiceFileURL   string `json:"voiceFileURL"`
-}
+//
+//type IOSVoIPPushAlarmInfoSQSMessage struct {
+//	AlarmID        string `json:"alarmID"`
+//	UserID         string `json:"userID"`
+//	SNSEndpointArn string `json:"snsEndpointArn"`
+//	CharaID        string `json:"charaID"`
+//	CharaName      string `json:"charaName"`
+//	VoiceFileURL   string `json:"voiceFileURL"`
+//}
