--- application/api/repository/environment.go
@@ -1,5 +1,0 @@
-package repository
-
-type Environment struct {
-	ResourceBaseURL string
-}
