+++ application/api/util/validator/chara_validator.go
@@ -2,11 +2,11 @@ package validator
 
 import (
 	"errors"
-	"github.com/takoikatakotako/charalarm/api/entity/database"
 	"github.com/takoikatakotako/charalarm/api/util/message"
+	"github.com/takoikatakotako/charalarm/repository/entity"
 )
 
-func ValidateChara(chara database.Chara) error {
+func ValidateChara(chara entity.Chara) error {
 	// CharaID
 	if !IsValidUUID(chara.CharaID) {
 		return errors.New(message.ErrorInvalidValue + ": CharaID")
