+++ application/api/util/validator/user_validator.go
@@ -2,12 +2,12 @@ package validator
 
 import (
 	"errors"
-	"github.com/takoikatakotako/charalarm/api/entity/database"
 	"github.com/takoikatakotako/charalarm/api/util/message"
+	"github.com/takoikatakotako/charalarm/repository/entity"
 	"time"
 )
 
-func ValidateUser(user database.User) error {
+func ValidateUser(user entity.User) error {
 	// UserID
 	if !IsValidUUID(user.UserID) {
 		return errors.New(message.ErrorInvalidValue + ": UserID")
@@ -50,7 +50,7 @@ func ValidateUser(user database.User) error {
 	return ValidateUserIOSPlatformInfo(user.IOSPlatformInfo)
 }
 
-func ValidateUserIOSPlatformInfo(userIOSPlatformInfo database.UserIOSPlatformInfo) error {
+func ValidateUserIOSPlatformInfo(userIOSPlatformInfo entity.UserIOSPlatformInfo) error {
 	// PushTokenが空文字の場合はPushTokenSNSEndpointも空文字
 	if userIOSPlatformInfo.PushToken == "" && userIOSPlatformInfo.PushTokenSNSEndpoint != "" {
 		return errors.New(message.ErrorInvalidValue + ": PushToken or PushTokenSNSEndpoint")
