+++ application/repository/aws_sns.go
@@ -7,6 +7,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/sns"
 	"github.com/takoikatakotako/charalarm/repository/entity"
+	"strings"
 )
 
 const (
@@ -83,4 +84,86 @@ func (a *AWS) CheckPlatformEndpointEnabled(endpoint string) error {
 		return errors.New("EndpointがFalse")
 	}
 	return nil
+}
+
+// SNSDeletePlatformApplicationEndpoint エンドポイントを削除するコードを追加
+func (a *AWS) SNSDeletePlatformApplicationEndpoint(endpointArn string) error {
+	client, err := a.createSNSClient()
+	if err != nil {
+		return err
+	}
+
+	// プッシュ通知を発火
+	input := &sns.DeleteEndpointInput{
+		EndpointArn: aws.String(endpointArn),
+	}
+
+	_, err = client.DeleteEndpoint(context.Background(), input)
+	return err
+}
+
+// CreateIOSPushPlatformEndpoint iOS Platform Endpoint
+func (a *AWS) CreateIOSPushPlatformEndpoint(pushToken string) (string, error) {
+	platformApplicationArn, err := a.getPlatformApplicationARN(iOSPushPlatformApplication)
+	if err != nil {
+		return "", err
+	}
+	return a.createPlatformEndpoint(platformApplicationArn, pushToken)
+}
+
+// CreateIOSVoipPushPlatformEndpoint iOS Platform Endpoint
+func (a *AWS) CreateIOSVoipPushPlatformEndpoint(pushToken string) (string, error) {
+	platformApplicationArn, err := a.getPlatformApplicationARN(iOSVoIPPushPlatformApplication)
+	if err != nil {
+		return "", err
+	}
+	return a.createPlatformEndpoint(platformApplicationArn, pushToken)
+}
+
+//////////////////////////////
+// Private Methods
+//////////////////////////////
+
+// getQueueURL QueueのURLを取得する
+func (a *AWS) getPlatformApplicationARN(queueName string) (string, error) {
+	// SQSClient作成
+	client, err := a.createSNSClient()
+	if err != nil {
+		return "", err
+	}
+
+	// PlatformApplication を取得
+	input := &sns.ListPlatformApplicationsInput{}
+	output, err := client.ListPlatformApplications(context.Background(), input)
+	if err != nil {
+		return "", err
+	}
+
+	for _, platformApplication := range output.PlatformApplications {
+		platformApplicationArn := *platformApplication.PlatformApplicationArn
+		if strings.Contains(platformApplicationArn, queueName) {
+			return platformApplicationArn, nil
+		}
+	}
+
+	return "", errors.New("platform Application Not Found")
+}
+
+func (a *AWS) createPlatformEndpoint(platformApplicationArn string, pushToken string) (string, error) {
+	client, err := a.createSNSClient()
+	if err != nil {
+		return "", err
+	}
+
+	// エンドポイント作成
+	getInput := &sns.CreatePlatformEndpointInput{
+		PlatformApplicationArn: aws.String(platformApplicationArn),
+		Token:                  aws.String(pushToken),
+	}
+	result, err := client.CreatePlatformEndpoint(context.Background(), getInput)
+	if err != nil {
+		return "", err
+	}
+
+	return *result.EndpointArn, nil
 }
